+++ docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # ChangeLog
 
+# Version 1.0.0-alpha91 (2023-04-28)
+- Improves payments error handling
+- Fixes crashes
+- Improves Dashboard loading
+
+# Version 1.0.0-alpha90 (2023-04-14)
+- Fixes recommendations issues
+
+# Version 1.0.0-alpha89 (2023-04-13)
+- Adds sharing inside Sphinx
+
 # Version 1.0.0-alpha88 (2023-04-10)
 - Adds feed item description view